refactor(ui): upgrade Angular to version 20.1 (#101)
* chore(deps-dev): update @angular/cli to v18.2.20

This commit updates the `@angular/cli` dependency from version 17.3.1 to version 18.2.20. This is a development dependency used for building and scaffolding the Angular application. The update includes bug fixes, performance improvements, and new features.

* feat: upgrade to Angular 18

This commit upgrades the Angular dependencies to version 18.2.13 and Angular CLI to 18.2.20. It also updates the `app.module.ts` to use `provideHttpClient` with `withInterceptorsFromDi` for providing the HttpClient. This is the recommended approach in Angular 18 for enabling dependency injection of interceptors.

* chore(deps): update @angular/material to v18.2.14

This commit updates the `@angular/material` dependency from version 17.3.1 to version 18.2.14. This includes bug fixes, performance improvements, and new features.

* chore(deps-dev): update @angular/cli to v19.2.15

This commit updates the `@angular/cli` dependency from version 18.2.20 to version 19.2.15. This is a development dependency used for building and scaffolding the Angular application. The update includes bug fixes, performance improvements, and new features.

* feat: upgrade to Angular 20

This commit upgrades the Angular dependencies to version 20.1.0 and Angular CLI to 20.1.0. It also updates the templates in `create-deployment` and `edit-deployment` components to use the new `@` syntax for control flow, replacing the `*ngIf` and `*ngFor` directives. This change improves the readability and performance of the templates.

* chore(deps): update angular material and cdk to v19.2.19 and cli to v20.1.0

This commit updates the following dependencies:
- `@angular/material` from version 18.2.14 to version 19.2.19
- `@angular/cdk` from version 18.2.14 to version 19.2.19
- `@angular/cli` from version 19.2.15 to version 20.1.0

These updates include bug fixes, performance improvements, and new features.

* chore(deps): update angular material and cdk to v20.1.0

This commit updates the following dependencies:
- `@angular/material` from version 19.2.19 to version 20.1.0
- `@angular/cdk` from version 19.2.19 to version 20.1.0

These updates include bug fixes, performance improvements, and new features.

* refactor(ui): migrate to standalone components and routes

This commit migrates the Angular application to use standalone components, replacing the traditional NgModule-based approach. This simplifies the application structure and improves load times by eliminating unnecessary modules.

Key changes include:

- Removal of `AppModule` and `AppRoutingModule`.
- Creation of `app.routes.ts` to define application routes.
- Bootstraping `AppComponent` directly in `main.ts` using `bootstrapApplication` and providing routes and other providers.
- Conversion of `AppComponent`, `CreateDeploymentComponent`, `DeploymentDashboardComponent`, and `EditDeploymentComponent` to standalone components by adding the `standalone: true` flag and importing required modules directly into the component.
- Updating component spec files to use `imports` instead of `declarations`.
- Disabling analytics in `angular.json`.

* refactor(ui): used prettier to apply minor template and style adjustments

This commit applies minor template and style adjustments to improve the UI.

Key changes include:

- Corrected indentation and spacing in multiple component templates for better readability.
- Updated the placement of the closing tag for the `mat-icon` element in `create-deployment.component.html` and `edit-deployment.component.html`.
- Removed unnecessary blank lines in `app.component.ts`.
- Updated imports to be multiline for readability.